refactor thunks and add types
