+++ protocol-designer/src/steplist/actions/actions.js
@@ -1,5 +1,6 @@
 // @flow
 import { getOrderedStepIds } from '../../step-forms/selectors'
+import { getNextNonTerminalItemStepId } from '../utils'
 import type { GetState, ThunkAction, ThunkDispatch } from '../../types'
 import type { StepIdType, FormData } from '../../form-types'
 import type { ChangeFormPayload } from './types'
@@ -7,7 +8,6 @@ import type {
   ClearSelectedItemAction,
   SelectMultipleStepsAction,
 } from '../../ui/steps'
-
 export type ChangeSavedStepFormAction = {|
   type: 'CHANGE_SAVED_STEP_FORM',
   payload: ChangeFormPayload,
@@ -50,7 +50,11 @@ export type DeleteMultipleStepsAction = {|
 // delete the steps, then update the selection
 export const deleteMultipleSteps = (
   stepIds: Array<StepIdType>
-): ThunkAction<*> => (dispatch: ThunkDispatch<*>, getState: GetState) => {
+): ThunkAction<
+  | DeleteMultipleStepsAction
+  | ClearSelectedItemAction
+  | SelectMultipleStepsAction
+> => (dispatch: ThunkDispatch<*>, getState: GetState) => {
   const orderedStepIds = getOrderedStepIds(getState())
   const deleteMultipleStepsAction: DeleteMultipleStepsAction = {
     type: 'DELETE_MULTIPLE_STEPS',
@@ -64,20 +68,7 @@ export const deleteMultipleSteps = (
     }
     dispatch(clearSelectedItemAction)
   } else {
-    // select the next non terminal item in multi select mode
-    let highestDeletedIndex = stepIds.reduce((highestIndex, val) => {
-      const currentStepIndex = orderedStepIds.indexOf(val)
-      return Math.max(currentStepIndex, highestIndex)
-    }, 0)
-    let nextStepId = orderedStepIds[highestDeletedIndex + 1]
-    while (!nextStepId) {
-      highestDeletedIndex -= 1
-      const potentialNextStepId = orderedStepIds[highestDeletedIndex]
-      if (stepIds.includes(potentialNextStepId)) {
-        continue
-      }
-      nextStepId = potentialNextStepId
-    }
+    const nextStepId = getNextNonTerminalItemStepId(orderedStepIds, stepIds)
     const selectMultipleStepsAction: SelectMultipleStepsAction = {
       type: 'SELECT_MULTIPLE_STEPS',
       payload: { stepIds: [nextStepId], lastSelected: nextStepId },
